% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drill.R
\name{dim_hierarchy}
\alias{dim_hierarchy}
\title{Ordered list to create a  hierarchical dimension}
\usage{
dim_hierarchy(...)
}
\arguments{
\item{...}{Ordered set of measures names}
}
\description{
Ordered list to create a  hierarchical dimension
}
\examples{

sales \%>\%
  rows(date = dim_hierarchy(year_id, month_id)) \%>\%
  values(n())
}
