% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drill.R
\name{drill}
\alias{drill}
\title{Drill into a hierarchy dimension}
\usage{
drill(.data, ...)
}
\arguments{
\item{.data}{A data.frame or a pivot_prep object}

\item{...}{Variable or variables to drill by}
}
\description{
Drill into a hierarchy dimension
}
\examples{

retail_orders \%>\%
  rows(order_date = dim_hierarchy(
    year = as.integer(format(orderdate, "\%Y")),
    month = as.integer(format(orderdate, "\%m"))
  )) \%>\%
  values(sum(sales))

}
