% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot.R
\name{pivot}
\alias{pivot}
\title{Switch the columns with the rows on a pivot table}
\usage{
pivot(.data)
}
\arguments{
\item{.data}{A pivot_table object}
}
\description{
Switch the columns with the rows on a pivot table
}
\examples{
sales \%>\%
  rows(status) \%>\%
  columns(year_id) \%>\%
  values(n()) \%>\%
  pivot()
}
