% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot-default.R
\name{pivot_default_totals}
\alias{pivot_default_totals}
\title{Set a default to display totals for an R session}
\usage{
pivot_default_totals(include_column_totals = TRUE,
  include_row_totals = TRUE)
}
\arguments{
\item{include_column_totals}{Indicates if the column totals are included in the pivot table}

\item{include_row_totals}{Indicates if the row totals are included in the pivot table}
}
\description{
Set a default to display totals for an R session
}
\examples{

pivot_default_totals(TRUE, TRUE)

retail_orders \%>\%
  pivot_rows(status, country) \%>\%
  pivot_values(n())

}
