% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot-default.R
\name{pivot_default_values}
\alias{pivot_default_values}
\title{Set the default agregation for an R session}
\usage{
pivot_default_values(...)
}
\arguments{
\item{...}{Variables or calculation to aggregate by}
}
\description{
Set the default agregation for an R session
}
\details{
It records the named quotation variables into via the options() command, inside
the "pivotable_default_values" option.
}
\examples{

pivot_default_values(n())

retail_orders \%>\%
  pivot_rows(status)

pivot_default_values(sum(sales))

retail_orders \%>\%
  pivot_rows(status)
}
