% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_drill.R
\name{pivot_drill}
\alias{pivot_drill}
\title{Drill into a hierarchy dimension}
\usage{
pivot_drill(.data, ...)
}
\arguments{
\item{.data}{A data.frame or a pivot_prep object}

\item{...}{Variable or variables to drill by}
}
\description{
Drill into a hierarchy dimension
}
\examples{

retail_orders \%>\%
  pivot_rows(order_date = dim_hierarchy(
    year = as.integer(format(orderdate, "\%Y")),
    month = as.integer(format(orderdate, "\%m"))
  )) \%>\%
  pivot_values(sum(sales)) \%>\%
  pivot_drill(order_date)

}
