% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_flip.R
\name{pivot_flip}
\alias{pivot_flip}
\title{Switch the columns with the rows on a pivot table}
\usage{
pivot_flip(.data)
}
\arguments{
\item{.data}{A pivot_table object}
}
\description{
Switch the columns with the rows on a pivot table
}
\examples{
retail_orders \%>\%
  pivot_rows(status) \%>\%
  pivot_columns(country) \%>\%
  pivot_values(n()) \%>\%
  pivot_flip()
}
