% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot-verbs.R
\name{pivot_focus}
\alias{pivot_focus}
\title{Filter a pivot table}
\usage{
pivot_focus(.data, ...)
}
\arguments{
\item{.data}{A data.frame or a pivot_prep object}

\item{...}{Variables or calculation to group by}
}
\description{
Filter a pivot table
}
\examples{

retail_orders \%>\%
  pivot_rows(status) \%>\%
  pivot_columns(country) \%>\%
  pivot_values(total_sales = sum(sales)) \%>\%
  pivot_focus(
    country \%in\% c("Japan", "USA", "UK"),
    status == "Shipped",
    total_sales > 200000
  )
}
