% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot-verbs.R
\name{pivot_totals}
\alias{pivot_totals}
\title{Control if the totals will display when printed}
\usage{
pivot_totals(.data, include_column_totals = TRUE,
  include_row_totals = TRUE)
}
\arguments{
\item{.data}{A data.frame or a pivot_prep object}

\item{include_column_totals}{Indicates if the column totals are included in the pivot table}

\item{include_row_totals}{Indicates if the row totals are included in the pivot table}
}
\description{
Control if the totals will display when printed
}
\examples{

retail_orders \%>\%
  pivot_rows(status, country) \%>\%
  pivot_values(n()) \%>\%
  pivot_totals(include_row_totals = FALSE)
}
