% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot-prep.R
\name{prep_measures}
\alias{prep_measures}
\title{Develop values to aggregate by for pivot tables}
\usage{
prep_measures(x, ...)
}
\arguments{
\item{x}{A pivot_prep object}

\item{...}{A set of variable or named variables}
}
\description{
Develop values to aggregate by for pivot tables
}
\examples{

sales_pivot <- retail_orders \%>\%
  prep_dimensions(order_date = dim_hierarchy(
    year = as.integer(format(orderdate, "\%Y")),
    month = as.integer(format(orderdate, "\%m"))
  )) \%>\%
  prep_measures(no_orders = n(), total_orders = sum(sales))

sales_pivot \%>\%
  pivot_columns(order_date) \%>\%
  pivot_values(total_orders)
}
