% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-sales.R
\docType{data}
\name{retail_orders}
\alias{retail_orders}
\title{Samples sales data}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 307 rows and 7 columns.}
\usage{
retail_orders
}
\description{
Samples sales data
}
\keyword{datasets}
