% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-sales.R
\docType{data}
\name{retail_orders_details}
\alias{retail_orders_details}
\title{Samples sales data}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2823 rows and 24 columns.}
\usage{
retail_orders_details
}
\description{
Samples sales data
}
\keyword{datasets}
