% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_verbs.R
\name{rows-cols}
\alias{rows-cols}
\alias{rows}
\alias{columns}
\alias{focus}
\title{Manipulate pivot tables}
\usage{
rows(.data, ...)

columns(.data, ...)

focus(.data, ...)
}
\arguments{
\item{.data}{A data.frame or a pivot_prep object}

\item{...}{Variables or calculation to group by}
}
\description{
Manipulate pivot tables
}
\examples{

retail_orders \%>\%
  rows(status) \%>\%
  columns(country) \%>\%
  values(n())
}
