% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-sales.R
\docType{data}
\name{sales}
\alias{sales}
\title{Samples sales data}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2823 rows and 24 columns.}
\usage{
sales
}
\description{
Samples sales data
}
\keyword{datasets}
