% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_cat.R
\name{set_cat}
\alias{set_cat}
\title{Internal function for creating pivot tables}
\usage{
set_cat(.data, atr = "", ...)
}
\arguments{
\item{.data}{data.frame, pivot_prep or pivot_table object}

\item{atr}{String value: columns, rows, or values}

\item{...}{Variables or named variables}
}
\description{
Internal function for creating pivot tables
}
