% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_prep.R
\name{start_pivot_prep}
\alias{start_pivot_prep}
\title{Initialize a pivot data object}
\usage{
start_pivot_prep(.data)
}
\arguments{
\item{.data}{A data.frame or tbl object}
}
\description{
Initialize a pivot data object
}
\examples{
sales_pivot <- sales \%>\%
  start_pivot_prep() \%>\%
  dimensions(order_date = dim_hierarchy(year_id, month_id)) \%>\%
  measures(no_orders = n(), total_orders = sum(sales))

sales_pivot \%>\%
  columns(order_date) \%>\%
  values(total_orders)

}
