% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_pivottabler.R
\name{to_pivottabler}
\alias{to_pivottabler}
\title{Coerce to a pivottabler object}
\usage{
to_pivottabler(x, include_column_totals = NULL,
  include_row_totals = NULL)
}
\arguments{
\item{x}{A pivot_prep or pivot_table object}

\item{include_column_totals}{Indicates if the column totals are included in the pivot table}

\item{include_row_totals}{Indicates if the row totals are included in the pivot table}
}
\description{
Coerce to a pivottabler object
}
\examples{

pt <- retail_orders \%>\%
  pivot_rows(status) \%>\%
  pivot_columns(country) \%>\%
  pivot_values(n()) \%>\%
  to_pivottabler()

pt$asMatrix()
}
