% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_verbs.R
\name{values}
\alias{values}
\title{Add an aggregation to a pivot table}
\usage{
values(.data, ...)
}
\arguments{
\item{.data}{A data.frame or a pivot_prep object}

\item{...}{Variables or calculation to aggregate by}
}
\description{
Add an aggregation to a pivot table
}
\examples{

sales \%>\%
  rows(status) \%>\%
  values(n())

}
