% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drill.R
\name{drill}
\alias{drill}
\title{Drill into a hierarchy dimension}
\usage{
drill(.data, ...)
}
\arguments{
\item{.data}{A data.frame or a pivot_prep object}

\item{...}{Variable or variables to drill by}
}
\description{
Drill into a hierarchy dimension
}
\examples{

sales \%>\%
  rows(date = dim_hierarchy(year_id, month_id)) \%>\%
  values(n()) \%>\%
  drill()

}
