% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_prep.R
\name{measures}
\alias{measures}
\title{Develop values to aggregate by for pivot tables}
\usage{
measures(x, ...)
}
\arguments{
\item{x}{A pivot_prep object}

\item{...}{A set of variable or named variables}
}
\description{
Develop values to aggregate by for pivot tables
}
\examples{
sales_pivot <- sales \%>\%
  dimensions(order_date = dim_hierarchy(year_id, month_id)) \%>\%
  measures(no_orders = n(), total_orders = sum(sales))

sales_pivot \%>\%
  columns(order_date) \%>\%
  values(total_orders)
}
