% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_pivottabler.R
\name{to_pivottabler}
\alias{to_pivottabler}
\title{Coerce to a pivottabler object}
\usage{
to_pivottabler(x)
}
\arguments{
\item{x}{A pivot_prep or pivot_table object}
}
\description{
Coerce to a pivottabler object
}
\examples{

pt <- retail_orders \%>\%
  rows(status) \%>\%
  columns(country) \%>\%
  values(n()) \%>\%
  to_pivottabler()

pt$asMatrix()
}
